third_party/WebKit/LayoutTests/sensor/idl-AmbientLightSensor.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
test(function() {
  // Test that AmbientLightSensor interface exists
  assert_true(new AmbientLightSensor() instanceof AmbientLightSensor);
}, 'AmbientLightSensor IDL test');
</script>
